api: Validate Set/Update requests using fieldmask
