+++ pkg/ttnpb/gateway.go
@@ -14,20 +14,16 @@
 
 package ttnpb
 
-import "context"
-
-// IsZero reports whether ids represent zero identifiers.
-func (ids ApplicationWebhookIdentifiers) IsZero() bool {
-	return ids.GetWebhookID() == "" &&
-		ids.GetApplicationID() == ""
-}
+import (
+	"context"
+)
 
 // ValidateContext wraps the generated validator with (optionally context-based) custom checks.
-func (m *SetApplicationWebhookRequest) ValidateContext(context.Context) error {
+func (m *UpdateGatewayRequest) ValidateContext(context.Context) error {
 	if len(m.FieldMask.Paths) == 0 {
 		return m.ValidateFields()
 	}
-	return m.ValidateFields(append(fieldsWithPrefix("webhook", m.FieldMask.Paths...),
-		"webhook.ids",
+	return m.ValidateFields(append(fieldsWithPrefix("gateway", m.FieldMask.Paths...),
+		"gateway.ids",
 	)...)
 }
